Use u64 for neighbour nr_cellid: is up to 38 bits

+++ openair2/RRC/NR/nr_rrc_defs.h
@@ -367,7 +367,7 @@ typedef struct {
 } nr_neighbour_gnb_configuration_t;
 
 typedef struct neighbour_cell_configuration_s {
-  int nr_cell_id;
+  uint64_t nr_cell_id;
   seq_arr_t *neighbour_cells;
 } neighbour_cell_configuration_t;
